Recommended package structure, new installation instructions, port to pengines package v2.0
