+++ .gitignore
@@ -1,4 +1,5 @@
 *.*~
 *~
 backup.js
-
+node_modules
+package-lock.json
