+++ package.json
@@ -1,8 +1,7 @@
 {
   "name": "node-red-contrib-pengines",
-  "version": "1.0.0",
-  "description": "node to query SWI-Prolog pengines servers",
-  "main": "index.js",
+  "version": "1.0.1",
+  "description": "Node-RED node to query SWI-Prolog pengines servers",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -19,10 +18,13 @@
   "bugs": {
     "url": "https://github.com/Anniepoo/node-red-pengines/issues"
   },
+  "dependencies": {
+    "pengines": "^2.0.0"
+  },
   "homepage": "https://github.com/Anniepoo/node-red-pengines",
-  "node-red" : {
-        "nodes": {
-            "pengine-rpc": "pengine-rpc.js"
-        }
+  "node-red": {
+    "nodes": {
+      "pengine-rpc": "pengine-rpc.js"
     }
+  }
 }
